refactor: 테스트 파일을 관련 함수와 같은 디렉토리로 이동
