+++ src/domains/shopping-cart/mocks/mockProductResponse.ts
