+++ src/domains/shopping-cart/mocks/mockShoppingCartResponse.ts
