+++ src/domains/shopping-cart/utils/getTotalPrice/getTotalPrice.ts
@@ -1,4 +1,4 @@
-import { CartItemTypes } from "../types/cartItem";
+import { CartItemTypes } from "../../types/cartItem";
 
 interface getTotalPriceProps {
   cartItems: CartItemTypes[];
