+++ test/setup.ts
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom";
 import "@testing-library/jest-dom/vitest";
-import { server } from "../src/domains/shopping-cart/__test__/mocks/server";
+import { server } from "../src/domains/shopping-cart/mocks/server";
 
 beforeAll(() => server.listen());
 afterAll(() => server.close());
